+++ clients/admin-ui/src/features/config-wizard/OrganizationInfoForm.tsx
@@ -26,7 +26,6 @@ import {
 const useOrganizationInfoForm = (handleChangeStep: Function) => {
   const [createOrganization] = useCreateOrganizationMutation();
   const [updateOrganization] = useUpdateOrganizationMutation();
-  // FUTURE TODO: Need a way to check for an existing fides_key from the start of the wizard
   const { data: existingOrg } = useGetOrganizationByFidesKeyQuery(
     "default_organization"
   );
@@ -42,8 +41,7 @@ const useOrganizationInfoForm = (handleChangeStep: Function) => {
         name: values.name ?? existingOrg?.name,
         description: values.description ?? existingOrg?.description,
         fides_key: existingOrg?.fides_key ?? "default_organization",
-        // FUTURE TODO: Need to check with this body that if they have a fides_key assigned,
-        // then assign that existing one
+        organization_fides_key: "default_organization",
       };
 
       setIsLoading(true);
@@ -66,10 +64,9 @@ const useOrganizationInfoForm = (handleChangeStep: Function) => {
             description: errorMsg,
           });
           return;
-        } 
-          toast.closeAll();
-          handleChangeStep(1);
-        
+        }
+        toast.closeAll();
+        handleChangeStep(1);
       } else {
         const updateOrganizationResult = await updateOrganization(
           organizationBody
@@ -88,10 +85,9 @@ const useOrganizationInfoForm = (handleChangeStep: Function) => {
             description: errorMsg,
           });
           return;
-        } 
-          toast.closeAll();
-          handleChangeStep(1);
-        
+        }
+        toast.closeAll();
+        handleChangeStep(1);
       }
 
       setIsLoading(false);