+++ clients/admin-ui/src/features/system/types.ts
@@ -6,7 +6,7 @@ import {
 
 export interface System extends FidesBase {
   registry_id?: number;
-  meta?: Record<string, string>;
+  meta?: { tags: string | undefined };
   fidesctl_meta?: SystemMetadata;
   system_type: string;
   data_responsibility_title: DataResponsibilityTitle;
@@ -26,7 +26,7 @@ interface SystemMetadata {
   endpoint_port?: string;
 }
 
-interface PrivacyDeclaration {
+export interface PrivacyDeclaration {
   name: string;
   data_categories: FidesKey[];
   data_use: FidesKey;